separate out code for slecting model name
