+++ ui/litellm-dashboard/src/components/model_dashboard.tsx
@@ -16,6 +16,7 @@ import {
   AccordionHeader,
   AccordionBody,
 } from "@tremor/react";
+import ConditionalPublicModelName from "./add_model/conditional_public_model_name";
 import {
   TabPanel,
   TabPanels,
@@ -1823,47 +1824,16 @@ const ModelDashboard: React.FC<ModelDashboardProps> = ({
 
                   {/* Conditionally Render "Public Model Name" */}
                   <Form.Item
-                    shouldUpdate={(prevValues, currentValues) => prevValues.model !== currentValues.model || prevValues.custom_llm_provider !== currentValues.custom_llm_provider}
+                    shouldUpdate={(prevValues, currentValues) => 
+                      prevValues.model !== currentValues.model || 
+                      prevValues.custom_llm_provider !== currentValues.custom_llm_provider
+                    }
                   >
                     {({ getFieldValue }) => {
                       const selectedModels = getFieldValue('model') || [];
                       const showPublicModelName = !selectedModels.includes('all-wildcard');
 
-                      return (
-                        <>
-                          {showPublicModelName && (
-                            <>
-                              <Form.Item
-                                label="Public Model Name"
-                                name="model_name"
-                                tooltip="Model name your users will pass in. Also used for load-balancing, LiteLLM will load balance between all models with this public name."
-                                className="mb-0"
-                                rules={[
-                                  ({ getFieldValue }) => ({
-                                    validator(_, value) {
-                                      const selectedModels = getFieldValue('model') || [];
-                                      if (!selectedModels.includes('all-wildcard') || value) {
-                                        return Promise.resolve();
-                                      }
-                                      return Promise.reject(new Error('Public Model Name is required unless "All Models" is selected.'));
-                                    },
-                                  }),
-                                ]}
-                              >
-                                <TextInput />
-                              </Form.Item>
-                              <Row>
-                                <Col span={10}></Col>
-                                <Col span={14}>
-                                  <Text className="mb-3 mt-1">
-                                    Model name your users will pass in.
-                                  </Text>
-                                </Col>
-                              </Row>
-                            </>
-                          )}
-                        </>
-                      );
+                      return <ConditionalPublicModelName show={showPublicModelName} />;
                     }}
                   </Form.Item>
                   <Form.Item
